Break up Leaflet components into separate files
